requirements.txt
streamlit>=1.28.0
plotly>=5.17.0
yfinance>=0.2.0
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
tensorflow>=2.15.0
TA-Lib>=0.4.0
psycopg2-binary>=2.9.0
